Generate a loading script and execute it with a bit of bootstrap code
This lets us unmap the old executable and also retains the ability to use
libc and keeps the old process workable as long as possible while writing
the loading script.

A few corner cases still remain, e.g. if anonymous mmap() ranges overlap
the load addresses of the new process, but the basic problem of the loader
overlapping the new process is fixed.